requirements.txt
Pillow
pytest
numpy
